requirements.txt
flask
chromadb
openai
pydantic
python-dotenv
tiktoken
gunicorn